requirements.txt
docker==7.1.0
Flask==3.0.3
Flask-Cors==4.0.1
mlflow==2.13.0
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.2
pillow==10.3.0
requests==2.32.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
torch==2.3.0
torchvision==0.18.0